requirements.txt
python-documentcloud
requests
python-dotenv
spacy==3.4.1
spacy-universal-sentence-encoder
https://github.com/MartinoMensio/spacy-universal-sentence-encoder/releases/download/v0.4.5/en_use_lg-0.4.5.tar.gz#en_use_lg-0.4.5
nltk
sklearn
numpy
annoy
